requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
pyvips~=2.2.1
numpy~=1.23.5
paquo~=0.6.0
torch~=1.13.0+cu116
torchvision
torchaudio
rasterio~=1.3.4
javabridge~=1.0.19
python-bioformats
Shapely~=1.8.5.post1
h5py~=3.7.0
future~=0.18.2
setuptools~=65.5.1